+++ yarn-project/end-to-end/src/e2e_p2p/p2p_network.ts
@@ -8,12 +8,12 @@ import { type Logger, createLogger } from '@aztec/foundation/log';
 import { ForwarderAbi, ForwarderBytecode, RollupAbi, TestERC20Abi } from '@aztec/l1-artifacts';
 import { SpamContract } from '@aztec/noir-contracts.js/Spam';
 import { type BootstrapNode } from '@aztec/p2p';
-import { createBootstrapNodeFromPrivateKey } from '@aztec/p2p/mocks';
 
 import getPort from 'get-port';
 import { getContract } from 'viem';
 import { privateKeyToAccount } from 'viem/accounts';
 
+import { createBootstrapNodeFromPrivateKey } from '../../../p2p/dest/test-helpers/index.js';
 import {
   ATTESTER_PRIVATE_KEYS_START_INDEX,
   PROPOSER_PRIVATE_KEYS_START_INDEX,
