+++ yarn-project/p2p/src/testbench/scripts/run_testbench.sh
@@ -1,0 +1,2 @@
+
+outputfile=$1
